+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## NAACCR XML Version History
 
+**Version 7.5**
+
+- Added logic to automatically move dateOfLastCancerStatus and dateOfLastCancerStatusFlag to each tumor when reading them at the Patient level.
+
 **Version 7.4**
 
 - Changed data level of dateOfLastCancerStatus and dateOfLastCancerStatusFlag from Patient to Tumor in N18 and N21 dictionaries.